refactor(contact): split http per slice
